+++ typed_sql/lib/src/typed_sql.expr_ext.dart
@@ -245,32 +245,11 @@ extension ExpressionNullableNum<T extends num> on Expr<T?> {
   /// {@macro orElse}
   Expr<T> orElseValue(T value) => orElse(toExpr(value));
 
-  /// {@template equals}
-  /// Compare this expression to [other] using `=` in SQL coalesced to `FALSE`.
-  ///
-  /// This is equivalent to `COALESCE(this = other, FALSE)` in SQL.
-  ///
-  /// The `.equals` _extension method_ requires that one of the two operators
-  /// are non-nullable. Because `NULL = NULL` evaluates to `UNKNOWN` in SQL,
-  /// which is surprising in a Dart context.
-  ///
-  /// If you wish to compare two _nullable expressions_ you can use:
-  ///  * [isNotDistinctFrom], to get `NULL` equivalent to `NULL`, or,
-  ///  * [equalsUnlessNull], to explicitely get the SQL `=` semantics.
-  /// {@endtemplate}
-  Expr<bool> equals(Expr<T> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  /// {@macro equals}
+  Expr<bool?> equals(Expr<T> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(T other) => equals(toExpr(other));
-
-  /// {@template notEquals}
-  /// Compare this expression to [other] using `<>` in SQL.
-  /// {@endtemplate}
-  Expr<bool> notEquals(Expr<T> other) => equals(other).not();
-
-  /// {@macro notEquals}
-  Expr<bool> notEqualsValue(T other) => notEquals(toExpr(other));
+  Expr<bool?> equalsValue(T other) => equals(toExpr(other));
 
   /// {@template isNotDistinctFrom}
   /// Compare this expression to [other] using `IS NOT DISTINCT FROM`.
@@ -284,8 +263,8 @@ extension ExpressionNullableNum<T extends num> on Expr<T?> {
   /// {@template equalsUnlessNull}
   /// Compare this expression to [other] using `=`.
   ///
-  /// This is equivalent to `this = other` in SQL, which if one of them is
-  /// `NULL` will return `NULL`.
+  /// This is equivalent to `this = other` in SQL, which if one of operands is
+  /// `NULL` will evaluate to `NULL`, following SQL three-valued logic.
   ///
   /// Use [isNotDistinctFrom] if you wish to compare expressions in a manner
   /// where `NULL` is considered equal to `NULL`.
@@ -316,17 +295,10 @@ extension ExpressionNullableString on Expr<String?> {
   Expr<String> orElseValue(String value) => orElse(toExpr(value));
 
   /// {@macro equals}
-  Expr<bool> equals(Expr<String> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<String> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(String other) => equals(toExpr(other));
-
-  /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<String> other) => equals(other).not();
-
-  /// {@macro notEquals}
-  Expr<bool> notEqualsValue(String other) => notEquals(toExpr(other));
+  Expr<bool?> equalsValue(String other) => equals(toExpr(other));
 
   /// {@macro isNotDistinctFrom}
   Expr<bool> isNotDistinctFrom(Expr<String?> other) =>
@@ -352,17 +324,10 @@ extension ExpressionNullableBool on Expr<bool?> {
   Expr<bool> orElseValue(bool value) => orElse(toExpr(value));
 
   /// {@macro equals}
-  Expr<bool> equals(Expr<bool> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<bool> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(bool other) => equals(toExpr(other));
-
-  /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<bool> other) => equals(other).not();
-
-  /// {@macro notEquals}
-  Expr<bool> notEqualsValue(bool other) => notEquals(toExpr(other));
+  Expr<bool?> equalsValue(bool other) => equals(toExpr(other));
 
   /// {@macro isNotDistinctFrom}
   Expr<bool> isNotDistinctFrom(Expr<bool?> other) =>
@@ -379,12 +344,12 @@ extension ExpressionNullableBool on Expr<bool?> {
   Expr<bool> isNotNull() => isNull().not();
 
   /// True, if this expression evaluates to `TRUE`.
-  Expr<bool> isTrue() => isNotDistinctFrom(Expr.true$);
+  Expr<bool> isTrue() => ExpressionIsTrue(this);
 
   /// True, if this expression evaluates to `FALSE`.
   ///
   /// If this is `NULL`, [isFalse] will evaluate to `FALSE`.
-  Expr<bool> isFalse() => isNotDistinctFrom(Expr.false$);
+  Expr<bool> isFalse() => ExpressionIsFalse(this);
 
   /// {@template and}
   /// Logical AND.
@@ -449,17 +414,10 @@ extension ExpressionNullableDateTime on Expr<DateTime?> {
   Expr<DateTime> orElseValue(DateTime value) => orElse(toExpr(value));
 
   /// {@macro equals}
-  Expr<bool> equals(Expr<DateTime> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<DateTime> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(DateTime other) => equals(toExpr(other));
-
-  /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<DateTime> other) => equals(other).not();
-
-  /// {@macro notEquals}
-  Expr<bool> notEqualsValue(DateTime other) => notEquals(toExpr(other));
+  Expr<bool?> equalsValue(DateTime other) => equals(toExpr(other));
 
   /// {@macro isNotDistinctFrom}
   Expr<bool> isNotDistinctFrom(Expr<DateTime?> other) =>
@@ -486,17 +444,10 @@ extension ExpressionNullableUint8List on Expr<Uint8List?> {
   Expr<Uint8List> orElseValue(Uint8List value) => orElse(toExpr(value));
 
   /// {@macro equals}
-  Expr<bool> equals(Expr<Uint8List> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<Uint8List> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(Uint8List other) => equals(toExpr(other));
-
-  /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<Uint8List> other) => equals(other).not();
-
-  /// {@macro notEquals}
-  Expr<bool> notEqualsValue(Uint8List other) => notEquals(toExpr(other));
+  Expr<bool?> equalsValue(Uint8List other) => equals(toExpr(other));
 
   /// {@macro isNotDistinctFrom}
   Expr<bool> isNotDistinctFrom(Expr<Uint8List?> other) =>
@@ -711,18 +662,43 @@ extension ExpressionNullableJsonValue on Expr<JsonValue?> {
 
 /// Extension methods for [bool] expressions.
 extension ExpressionBool on Expr<bool> {
-  /// {@macro equals}
-  Expr<bool> equals(Expr<bool?> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  /// {@template equals}
+  /// Compare this expression to [other] using `=` in SQL.
+  ///
+  /// This is equivalent to `this = other` in SQL, which if one of the operands
+  /// is `NULL` will evaluate to `NULL`, following SQL three-valued logic.
+  ///
+  /// The `.equals` method avoids the surprising behavior where `NULL = NULL`
+  /// evaluates to `NULL` by not allowing comparison between two nullable
+  /// operands.
+  ///
+  /// If you wish to compare _nullable expressions_ you can use:
+  ///  * [isNotDistinctFrom], to get `NULL` equivalent to `NULL`, or,
+  ///  * [equalsUnlessNull], to explicitely get the SQL `=` semantics.
+  /// {@endtemplate}
+  Expr<bool?> equals(Expr<bool?> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(bool? other) => equals(toExpr(other));
+  Expr<bool?> equalsValue(bool? other) => equals(toExpr(other));
+
+  /// {@template notEquals}
+  /// Compare this expression to [other] using `<>` in SQL.
+  ///
+  /// This is equivalent to `this <> other` in SQL,
+  ///
+  /// The `.notEquals` method avoids the surprising behavior where
+  /// `anything <> NULL` evaluates to `NULL` by not allowing any operand to be
+  /// nullable.
+  ///
+  /// If you wish to compare _nullable expressions_ you can use:
+  ///  * [isNotDistinctFrom], to get `NULL` equivalent to `NULL`, or,
+  ///  * [equalsUnlessNull], to explicitely get the SQL `=` semantics.
+  /// {@endtemplate}
+  Expr<bool> notEquals(Expr<bool> other) =>
+      ExpressionNotEquals(this, other).asNotNull();
 
   /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<bool?> other) => equals(other).not();
-
-  /// {@macro notEquals}
-  Expr<bool> notEqualsValue(bool? other) => notEquals(toExpr(other));
+  Expr<bool> notEqualsValue(bool other) => notEquals(toExpr(other));
 
   /// {@macro isNull}
   ///
@@ -847,17 +823,17 @@ extension ExpressionBool on Expr<bool> {
 /// Extension methods for [String] expressions.
 extension ExpressionString on Expr<String> {
   /// {@macro equals}
-  Expr<bool> equals(Expr<String?> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<String?> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(String? other) => equals(toExpr(other));
+  Expr<bool?> equalsValue(String? other) => equals(toExpr(other));
 
   /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<String?> other) => equals(other).not();
+  Expr<bool> notEquals(Expr<String> other) =>
+      ExpressionNotEquals(this, other).asNotNull();
 
   /// {@macro notEquals}
-  Expr<bool> notEqualsValue(String? other) => notEquals(toExpr(other));
+  Expr<bool> notEqualsValue(String other) => notEquals(toExpr(other));
 
   /// {@macro isNull}
   ///
@@ -1260,17 +1236,17 @@ extension ExpressionDouble on Expr<double> {
 /// Extension methods for [int] and [double] expressions.
 extension ExpressionNum<T extends num> on Expr<T> {
   /// {@macro equals}
-  Expr<bool> equals(Expr<T?> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<T?> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(T? other) => equals(toExpr(other));
+  Expr<bool?> equalsValue(T? other) => equals(toExpr(other));
 
   /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<T?> other) => equals(other).not();
+  Expr<bool> notEquals(Expr<T> other) =>
+      ExpressionNotEquals(this, other).asNotNull();
 
   /// {@macro notEquals}
-  Expr<bool> notEqualsValue(T? other) => notEquals(toExpr(other));
+  Expr<bool> notEqualsValue(T other) => notEquals(toExpr(other));
 
   /// {@macro isNull}
   ///
@@ -1348,17 +1324,17 @@ extension ExpressionNum<T extends num> on Expr<T> {
 /// Extension methods for [DateTime] expressions.
 extension ExpressionDateTime on Expr<DateTime> {
   /// {@macro equals}
-  Expr<bool> equals(Expr<DateTime?> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<DateTime?> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(DateTime? other) => equals(toExpr(other));
+  Expr<bool?> equalsValue(DateTime? other) => equals(toExpr(other));
 
   /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<DateTime?> other) => equals(other).not();
+  Expr<bool> notEquals(Expr<DateTime> other) =>
+      ExpressionNotEquals(this, other).asNotNull();
 
   /// {@macro notEquals}
-  Expr<bool> notEqualsValue(DateTime? other) => notEquals(toExpr(other));
+  Expr<bool> notEqualsValue(DateTime other) => notEquals(toExpr(other));
 
   /// {@macro isNull}
   ///
@@ -1438,17 +1414,17 @@ extension ExpressionDateTime on Expr<DateTime> {
 /// Extension methods for [Uint8List] expressions.
 extension ExpressionUint8List on Expr<Uint8List> {
   /// {@macro equals}
-  Expr<bool> equals(Expr<Uint8List?> other) =>
-      ExpressionEquals(this, other).orElseValue(false);
+  Expr<bool?> equals(Expr<Uint8List?> other) => ExpressionEquals(this, other);
 
   /// {@macro equals}
-  Expr<bool> equalsValue(Uint8List? other) => equals(toExpr(other));
+  Expr<bool?> equalsValue(Uint8List? other) => equals(toExpr(other));
 
   /// {@macro notEquals}
-  Expr<bool> notEquals(Expr<Uint8List?> other) => equals(other).not();
+  Expr<bool> notEquals(Expr<Uint8List> other) =>
+      ExpressionNotEquals(this, other).asNotNull();
 
   /// {@macro notEquals}
-  Expr<bool> notEqualsValue(Uint8List? other) => notEquals(toExpr(other));
+  Expr<bool> notEqualsValue(Uint8List other) => notEquals(toExpr(other));
 
   /// {@macro isNull}
   ///