+++ typed_sql/test/typed_sql/documentation/model_documentation_test.g.dart
@@ -344,7 +344,7 @@ extension ExpressionAuthorExt on Expr<Author> {
   /// in this row.
   SubQuery<(Expr<Book>,)> get books => $ForGeneratedCode
       .subqueryTable(_$Book._$table)
-      .where((r) => r.authorId.equals(authorId));
+      .where((r) => r.authorId.equalsUnlessNull(authorId));
 
   /// Get [SubQuery] of rows from the `books` table which
   /// reference this row.
@@ -355,7 +355,7 @@ extension ExpressionAuthorExt on Expr<Author> {
   /// in this row.
   SubQuery<(Expr<Book>,)> get booksEditedBy => $ForGeneratedCode
       .subqueryTable(_$Book._$table)
-      .where((r) => r.editorId.equals(authorId));
+      .where((r) => r.editorId.equalsUnlessNull(authorId));
 
   /// Do a subquery lookup of the row from table
   /// `books` referenced in
@@ -366,7 +366,7 @@ extension ExpressionAuthorExt on Expr<Author> {
   /// is equal to [favoriteBookId], if any.
   Expr<Book?> get favoriteBook => $ForGeneratedCode
       .subqueryTable(_$Book._$table)
-      .where((r) => r.bookId.equals(favoriteBookId))
+      .where((r) => r.bookId.equalsUnlessNull(favoriteBookId))
       .first;
 }
 
@@ -397,7 +397,7 @@ extension ExpressionNullableAuthorExt on Expr<Author?> {
   /// If this row is `NULL` the subquery is always be empty.
   SubQuery<(Expr<Book>,)> get books => $ForGeneratedCode
       .subqueryTable(_$Book._$table)
-      .where((r) => r.authorId.equalsUnlessNull(authorId).asNotNull());
+      .where((r) => r.authorId.equalsUnlessNull(authorId));
 
   /// Get [SubQuery] of rows from the `books` table which
   /// reference this row.
@@ -410,7 +410,7 @@ extension ExpressionNullableAuthorExt on Expr<Author?> {
   /// If this row is `NULL` the subquery is always be empty.
   SubQuery<(Expr<Book>,)> get booksEditedBy => $ForGeneratedCode
       .subqueryTable(_$Book._$table)
-      .where((r) => r.editorId.equalsUnlessNull(authorId).asNotNull());
+      .where((r) => r.editorId.equalsUnlessNull(authorId));
 
   /// Do a subquery lookup of the row from table
   /// `books` referenced in
@@ -423,7 +423,7 @@ extension ExpressionNullableAuthorExt on Expr<Author?> {
   /// If this row is `NULL` the subquery is always return `NULL`.
   Expr<Book?> get favoriteBook => $ForGeneratedCode
       .subqueryTable(_$Book._$table)
-      .where((r) => r.bookId.equalsUnlessNull(favoriteBookId).asNotNull())
+      .where((r) => r.bookId.equalsUnlessNull(favoriteBookId))
       .first;
 
   /// Check if the row is not `NULL`.
@@ -448,19 +448,19 @@ extension InnerJoinAuthorBookExt on InnerJoin<(Expr<Author>,), (Expr<Book>,)> {
   ///
   /// This will match rows where [Author.authorId] = [Book.authorId].
   Query<(Expr<Author>, Expr<Book>)> usingAuthor() =>
-      on((a, b) => a.authorId.equals(b.authorId));
+      on((a, b) => a.authorId.equalsUnlessNull(b.authorId));
 
   /// Join using the `editor` _foreign key_.
   ///
   /// This will match rows where [Author.authorId] = [Book.editorId].
   Query<(Expr<Author>, Expr<Book>)> usingEditor() =>
-      on((a, b) => a.authorId.equals(b.editorId));
+      on((a, b) => a.authorId.equalsUnlessNull(b.editorId));
 
   /// Join using the `favoriteBook` _foreign key_.
   ///
   /// This will match rows where [Author.favoriteBookId] = [Book.bookId].
   Query<(Expr<Author>, Expr<Book>)> usingFavoriteBook() =>
-      on((a, b) => b.bookId.equals(a.favoriteBookId));
+      on((a, b) => b.bookId.equalsUnlessNull(a.favoriteBookId));
 }
 
 extension LeftJoinAuthorBookExt on LeftJoin<(Expr<Author>,), (Expr<Book>,)> {
@@ -468,19 +468,19 @@ extension LeftJoinAuthorBookExt on LeftJoin<(Expr<Author>,), (Expr<Book>,)> {
   ///
   /// This will match rows where [Author.authorId] = [Book.authorId].
   Query<(Expr<Author>, Expr<Book?>)> usingAuthor() =>
-      on((a, b) => a.authorId.equals(b.authorId));
+      on((a, b) => a.authorId.equalsUnlessNull(b.authorId));
 
   /// Join using the `editor` _foreign key_.
   ///
   /// This will match rows where [Author.authorId] = [Book.editorId].
   Query<(Expr<Author>, Expr<Book?>)> usingEditor() =>
-      on((a, b) => a.authorId.equals(b.editorId));
+      on((a, b) => a.authorId.equalsUnlessNull(b.editorId));
 
   /// Join using the `favoriteBook` _foreign key_.
   ///
   /// This will match rows where [Author.favoriteBookId] = [Book.bookId].
   Query<(Expr<Author>, Expr<Book?>)> usingFavoriteBook() =>
-      on((a, b) => b.bookId.equals(a.favoriteBookId));
+      on((a, b) => b.bookId.equalsUnlessNull(a.favoriteBookId));
 }
 
 extension RightJoinAuthorBookExt on RightJoin<(Expr<Author>,), (Expr<Book>,)> {
@@ -488,19 +488,19 @@ extension RightJoinAuthorBookExt on RightJoin<(Expr<Author>,), (Expr<Book>,)> {
   ///
   /// This will match rows where [Author.authorId] = [Book.authorId].
   Query<(Expr<Author?>, Expr<Book>)> usingAuthor() =>
-      on((a, b) => a.authorId.equals(b.authorId));
+      on((a, b) => a.authorId.equalsUnlessNull(b.authorId));
 
   /// Join using the `editor` _foreign key_.
   ///
   /// This will match rows where [Author.authorId] = [Book.editorId].
   Query<(Expr<Author?>, Expr<Book>)> usingEditor() =>
-      on((a, b) => a.authorId.equals(b.editorId));
+      on((a, b) => a.authorId.equalsUnlessNull(b.editorId));
 
   /// Join using the `favoriteBook` _foreign key_.
   ///
   /// This will match rows where [Author.favoriteBookId] = [Book.bookId].
   Query<(Expr<Author?>, Expr<Book>)> usingFavoriteBook() =>
-      on((a, b) => b.bookId.equals(a.favoriteBookId));
+      on((a, b) => b.bookId.equalsUnlessNull(a.favoriteBookId));
 }
 
 /// `Table<Author>` conflict targets for use with `.onConflict`.
@@ -1057,7 +1057,7 @@ extension ExpressionBookExt on Expr<Book> {
   /// in this row.
   SubQuery<(Expr<Author>,)> get favoritedBy => $ForGeneratedCode
       .subqueryTable(_$Author._$table)
-      .where((r) => r.favoriteBookId.equals(bookId));
+      .where((r) => r.favoriteBookId.equalsUnlessNull(bookId));
 
   /// Do a subquery lookup of the row from table
   /// `authors` referenced in
@@ -1068,7 +1068,7 @@ extension ExpressionBookExt on Expr<Book> {
   /// is equal to [authorId].
   Expr<Author> get author => $ForGeneratedCode
       .subqueryTable(_$Author._$table)
-      .where((r) => r.authorId.equals(authorId))
+      .where((r) => r.authorId.equalsUnlessNull(authorId))
       .first
       .asNotNull();
 
@@ -1081,7 +1081,7 @@ extension ExpressionBookExt on Expr<Book> {
   /// is equal to [editorId], if any.
   Expr<Author?> get editor => $ForGeneratedCode
       .subqueryTable(_$Author._$table)
-      .where((r) => r.authorId.equals(editorId))
+      .where((r) => r.authorId.equalsUnlessNull(editorId))
       .first;
 }
 
@@ -1113,7 +1113,7 @@ extension ExpressionNullableBookExt on Expr<Book?> {
   /// If this row is `NULL` the subquery is always be empty.
   SubQuery<(Expr<Author>,)> get favoritedBy => $ForGeneratedCode
       .subqueryTable(_$Author._$table)
-      .where((r) => r.favoriteBookId.equalsUnlessNull(bookId).asNotNull());
+      .where((r) => r.favoriteBookId.equalsUnlessNull(bookId));
 
   /// Do a subquery lookup of the row from table
   /// `authors` referenced in
@@ -1126,7 +1126,7 @@ extension ExpressionNullableBookExt on Expr<Book?> {
   /// If this row is `NULL` the subquery is always return `NULL`.
   Expr<Author?> get author => $ForGeneratedCode
       .subqueryTable(_$Author._$table)
-      .where((r) => r.authorId.equalsUnlessNull(authorId).asNotNull())
+      .where((r) => r.authorId.equalsUnlessNull(authorId))
       .first;
 
   /// Do a subquery lookup of the row from table
@@ -1140,7 +1140,7 @@ extension ExpressionNullableBookExt on Expr<Book?> {
   /// If this row is `NULL` the subquery is always return `NULL`.
   Expr<Author?> get editor => $ForGeneratedCode
       .subqueryTable(_$Author._$table)
-      .where((r) => r.authorId.equalsUnlessNull(editorId).asNotNull())
+      .where((r) => r.authorId.equalsUnlessNull(editorId))
       .first;
 
   /// Check if the row is not `NULL`.
@@ -1165,19 +1165,19 @@ extension InnerJoinBookAuthorExt on InnerJoin<(Expr<Book>,), (Expr<Author>,)> {
   ///
   /// This will match rows where [Book.bookId] = [Author.favoriteBookId].
   Query<(Expr<Book>, Expr<Author>)> usingFavoriteBook() =>
-      on((a, b) => a.bookId.equals(b.favoriteBookId));
+      on((a, b) => a.bookId.equalsUnlessNull(b.favoriteBookId));
 
   /// Join using the `author` _foreign key_.
   ///
   /// This will match rows where [Book.authorId] = [Author.authorId].
   Query<(Expr<Book>, Expr<Author>)> usingAuthor() =>
-      on((a, b) => b.authorId.equals(a.authorId));
+      on((a, b) => b.authorId.equalsUnlessNull(a.authorId));
 
   /// Join using the `editor` _foreign key_.
   ///
   /// This will match rows where [Book.editorId] = [Author.authorId].
   Query<(Expr<Book>, Expr<Author>)> usingEditor() =>
-      on((a, b) => b.authorId.equals(a.editorId));
+      on((a, b) => b.authorId.equalsUnlessNull(a.editorId));
 }
 
 extension LeftJoinBookAuthorExt on LeftJoin<(Expr<Book>,), (Expr<Author>,)> {
@@ -1185,19 +1185,19 @@ extension LeftJoinBookAuthorExt on LeftJoin<(Expr<Book>,), (Expr<Author>,)> {
   ///
   /// This will match rows where [Book.bookId] = [Author.favoriteBookId].
   Query<(Expr<Book>, Expr<Author?>)> usingFavoriteBook() =>
-      on((a, b) => a.bookId.equals(b.favoriteBookId));
+      on((a, b) => a.bookId.equalsUnlessNull(b.favoriteBookId));
 
   /// Join using the `author` _foreign key_.
   ///
   /// This will match rows where [Book.authorId] = [Author.authorId].
   Query<(Expr<Book>, Expr<Author?>)> usingAuthor() =>
-      on((a, b) => b.authorId.equals(a.authorId));
+      on((a, b) => b.authorId.equalsUnlessNull(a.authorId));
 
   /// Join using the `editor` _foreign key_.
   ///
   /// This will match rows where [Book.editorId] = [Author.authorId].
   Query<(Expr<Book>, Expr<Author?>)> usingEditor() =>
-      on((a, b) => b.authorId.equals(a.editorId));
+      on((a, b) => b.authorId.equalsUnlessNull(a.editorId));
 }
 
 extension RightJoinBookAuthorExt on RightJoin<(Expr<Book>,), (Expr<Author>,)> {
@@ -1205,19 +1205,19 @@ extension RightJoinBookAuthorExt on RightJoin<(Expr<Book>,), (Expr<Author>,)> {
   ///
   /// This will match rows where [Book.bookId] = [Author.favoriteBookId].
   Query<(Expr<Book?>, Expr<Author>)> usingFavoriteBook() =>
-      on((a, b) => a.bookId.equals(b.favoriteBookId));
+      on((a, b) => a.bookId.equalsUnlessNull(b.favoriteBookId));
 
   /// Join using the `author` _foreign key_.
   ///
   /// This will match rows where [Book.authorId] = [Author.authorId].
   Query<(Expr<Book?>, Expr<Author>)> usingAuthor() =>
-      on((a, b) => b.authorId.equals(a.authorId));
+      on((a, b) => b.authorId.equalsUnlessNull(a.authorId));
 
   /// Join using the `editor` _foreign key_.
   ///
   /// This will match rows where [Book.editorId] = [Author.authorId].
   Query<(Expr<Book?>, Expr<Author>)> usingEditor() =>
-      on((a, b) => b.authorId.equals(a.editorId));
+      on((a, b) => b.authorId.equalsUnlessNull(a.editorId));
 }
 
 /// `Table<Book>` conflict targets for use with `.onConflict`.