+++ typed_sql/test/typed_sql/expr/datetime_test.dart
@@ -54,12 +54,12 @@ final _cases = [
   (
     name: 'epoch.equals(null)',
     expr: toExpr(epoch).equals(toExpr(null)),
-    expected: false,
+    expected: null,
   ),
   (
     name: 'today.equals(null)',
     expr: toExpr(today).equals(toExpr(null)),
-    expected: false,
+    expected: null,
   ),
 
   // Test for .equalsValue
@@ -96,12 +96,12 @@ final _cases = [
   (
     name: 'epoch.equalsValue(null)',
     expr: toExpr(epoch).equalsValue(null),
-    expected: false,
+    expected: null,
   ),
   (
     name: 'today.equalsValue(null)',
     expr: toExpr(today).equalsValue(null),
-    expected: false,
+    expected: null,
   ),
 
   // Test for .notEquals
@@ -137,16 +137,6 @@ final _cases = [
     expr: toExpr(epoch).notEquals(toExpr(today)),
     expected: true,
   ),
-  (
-    name: 'epoch.notEquals(null)',
-    expr: toExpr(epoch).notEquals(toExpr(null)),
-    expected: true,
-  ),
-  (
-    name: 'today.notEquals(null)',
-    expr: toExpr(today).notEquals(toExpr(null)),
-    expected: true,
-  ),
 
   // Test for .notEqualsValue
   (
@@ -177,16 +167,6 @@ final _cases = [
   (
     name: 'epoch.notEqualsValue(today)',
     expr: toExpr(epoch).notEqualsValue(today),
-    expected: true,
-  ),
-  (
-    name: 'epoch.notEqualsValue(null)',
-    expr: toExpr(epoch).notEqualsValue(null),
-    expected: true,
-  ),
-  (
-    name: 'today.notEqualsValue(null)',
-    expr: toExpr(today).notEqualsValue(null),
     expected: true,
   ),
 
@@ -481,7 +461,7 @@ void main() {
       final result = await db.select(
         (c.expr,),
       ).fetch();
-      check(result).isNotNull().equals(c.expected);
+      check(result).equals(c.expected);
     });
   }