+++ typed_sql/test/typed_sql/expr/nullable_int_test.dart
@@ -74,7 +74,7 @@ final _cases =
       (
         name: 'null.asInt().equals(42)',
         expr: toExpr(null).asInt().equals(toExpr(42)),
-        expected: false,
+        expected: null,
       ),
       (
         name: '42.equals(42)',
@@ -84,7 +84,7 @@ final _cases =
       (
         name: 'null.asInt().equals(0)',
         expr: toExpr(null).asInt().equals(toExpr(0)),
-        expected: false,
+        expected: null,
       ),
       (
         name: '0.equals(0)',
@@ -163,32 +163,40 @@ final _cases =
         expr: toExpr(0 as int?).equalsUnlessNull(toExpr(0)),
         expected: true,
       ),
-      // Expr<int?>.notEquals
+      // Expr<int?>.equalsUnlessNull(..).orElseValue(false).not()
       (
-        name: 'null.asInt().notEquals(42)',
-        expr: toExpr(null).asInt().notEquals(toExpr(42)),
+        name: 'null.asInt().equalsUnlessNull(42).orElseValue(false).not()',
+        expr: toExpr(
+          null,
+        ).asInt().equalsUnlessNull(toExpr(42)).orElseValue(false).not(),
         expected: true,
       ),
       (
-        name: '42.notEquals(42)',
-        expr: toExpr(42 as int?).notEquals(toExpr(42)),
+        name: '42.equalsUnlessNull(42).orElseValue(false).not()',
+        expr: toExpr(
+          42 as int?,
+        ).equalsUnlessNull(toExpr(42)).orElseValue(false).not(),
         expected: false,
       ),
       (
-        name: 'null.asInt().notEquals(0)',
-        expr: toExpr(null).asInt().notEquals(toExpr(0)),
+        name: 'null.asInt().equalsUnlessNull(0).orElseValue(false).not()',
+        expr: toExpr(
+          null,
+        ).asInt().equalsUnlessNull(toExpr(0)).orElseValue(false).not(),
         expected: true,
       ),
       (
-        name: '0.notEquals(0)',
-        expr: toExpr(0 as int?).notEquals(toExpr(0)),
+        name: '0.equalsUnlessNull(0).orElseValue(false).not()',
+        expr: toExpr(
+          0 as int?,
+        ).equalsUnlessNull(toExpr(0)).orElseValue(false).not(),
         expected: false,
       ),
       // Expr<int?>.equalsValue
       (
         name: 'null.asInt().equalsValue(42)',
         expr: toExpr(null).asInt().equalsValue(42),
-        expected: false,
+        expected: null,
       ),
       (
         name: '42.equalsValue(42)',
@@ -198,34 +206,14 @@ final _cases =
       (
         name: 'null.asInt().equalsValue(0)',
         expr: toExpr(null).asInt().equalsValue(0),
-        expected: false,
+        expected: null,
       ),
       (
         name: '0.equalsValue(0)',
         expr: toExpr(0 as int?).equalsValue(0),
         expected: true,
       ),
-      // Expr<int?>.notEqualsValue
-      (
-        name: 'null.asInt().notEqualsValue(42)',
-        expr: toExpr(null).asInt().notEqualsValue(42),
-        expected: true,
-      ),
-      (
-        name: '42.notEqualsValue(42)',
-        expr: toExpr(42 as int?).notEqualsValue(42),
-        expected: false,
-      ),
-      (
-        name: 'null.asInt().notEqualsValue(0)',
-        expr: toExpr(null).asInt().notEqualsValue(0),
-        expected: true,
-      ),
-      (
-        name: '0.notEqualsValue(0)',
-        expr: toExpr(0 as int?).notEqualsValue(0),
-        expected: false,
-      ),
+
       // Expr<int?>.isNull()
       (
         name: 'null.isNull()',